// SDKs/googleapis.com/identitytoolkit/v2/typescript/src/sdk/models/shared/googlecloudidentitytoolkitadminv2hashconfig.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";


export enum GoogleCloudIdentitytoolkitAdminV2HashConfigAlgorithmEnum {
    HashAlgorithmUnspecified = "HASH_ALGORITHM_UNSPECIFIED",
    HmacSha256 = "HMAC_SHA256",
    HmacSha1 = "HMAC_SHA1",
    HmacMd5 = "HMAC_MD5",
    Scrypt = "SCRYPT",
    PbkdfSha1 = "PBKDF_SHA1",
    Md5 = "MD5",
    HmacSha512 = "HMAC_SHA512",
    Sha1 = "SHA1",
    Bcrypt = "BCRYPT",
    Pbkdf2Sha256 = "PBKDF2_SHA256",
    Sha256 = "SHA256",
    Sha512 = "SHA512",
    StandardScrypt = "STANDARD_SCRYPT"
}


// GoogleCloudIdentitytoolkitAdminV2HashConfig
/** 
 * History information of the hash algorithm and key. Different accounts' passwords may be generated by different version.
**/
export class GoogleCloudIdentitytoolkitAdminV2HashConfig extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=algorithm" })
  algorithm?: GoogleCloudIdentitytoolkitAdminV2HashConfigAlgorithmEnum;

  @SpeakeasyMetadata({ data: "json, name=memoryCost" })
  memoryCost?: number;

  @SpeakeasyMetadata({ data: "json, name=rounds" })
  rounds?: number;

  @SpeakeasyMetadata({ data: "json, name=saltSeparator" })
  saltSeparator?: string;

  @SpeakeasyMetadata({ data: "json, name=signerKey" })
  signerKey?: string;
}